change font sizing to work on linux
When running on linux, specifying the font along with
the emoji characters results in blank characters being shown.
I suspect this is because the character is not part of
the requested font (Arial), but fallback to a font that
does support the character does not happen.

Really the font was specified so as to adjust the size
rather than the font name itself, so we also adjust things
so we're using a single font size for all characters for now.
Ideally we would be using a smaller font size for the numbers,
but we can adjust that later.

We choose a height=1 and width=2 throughout to achieve a
square rather than a rectangular grid.

Useful links when looking to adjust font and label sizes in future are:
https://stackoverflow.com/a/44920393/4421
https://www.delftstack.com/howto/python-tkinter/how-to-change-the-tkinter-label-font-size/
http://effbot.org/tkinterbook/label.htm