fix bombs left when avoiding initial bomb
If we needed to recreate the bombfield after the first click was a bomb,
we previously did not reset the bomb_left counter.

Here we move initialization of bombs_left to create_bombfield()
to ensure it's alway initialized.

We also only show the bombs remaining after the first click,
so that there is no confusing change in number after the first click