make controls consistent across systems
Button-2 is middle click on Linux or Windows systems.
Let's map this to standard click for now.
On Mac systems this is right click.

Button-3 is right click on Linux or Windows systems.

Also we map Control Click to right click,
as that's a standard way to get right click
on older Mac systems that don't support
two finger click to get the secondary click.